backend/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
slowapi
redis
pydantic-settings
python-multipart
